daohang/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
browsermob-proxy==0.8.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
furl==2.1.2
idna==2.10
iniconfig==1.1.1
orderedmultidict==1.0.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-assume==2.4.3
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
